Implemented CancellableFuture
