+++ src/main/java/com/hierynomus/smbj/connection/Request.java
@@ -16,19 +16,13 @@
 package com.hierynomus.smbj.connection;
 
 import com.hierynomus.mssmb2.SMB2Packet;
+import com.hierynomus.protocol.commons.concurrent.AFuture;
+import com.hierynomus.protocol.commons.concurrent.CancellableFuture;
 import com.hierynomus.protocol.commons.concurrent.Promise;
 import com.hierynomus.smbj.common.SMBRuntimeException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.util.Date;
 import java.util.UUID;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.Future;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.locks.ReentrantReadWriteLock;
 
 class Request {
 
@@ -67,64 +61,10 @@ class Request {
         return messageId;
     }
 
-    <T extends SMB2Packet> Future<T> getFuture(final CancelCallback callback) {
-        return new Future<T>() {
-            private final Logger logger = LoggerFactory.getLogger(Request.class);
-            private final AtomicBoolean cancelled = new AtomicBoolean(false);
-            private final ReentrantReadWriteLock lock = new ReentrantReadWriteLock();
+    <T extends SMB2Packet> AFuture<T> getFuture(final CancellableFuture.CancelCallback callback) {
+        //noinspection unchecked
+        return (AFuture<T>) new CancellableFuture<>(promise.future(), callback);
 
-            @Override
-            public boolean cancel(boolean mayInterruptIfRunning) {
-                lock.writeLock().lock();
-                try {
-                    if (isDone() || cancelled.getAndSet(true)) {
-                        // Already done or cancelled
-                        return false;
-                    } else {
-                        callback.cancel(messageId);
-                        return true;
-                    }
-                } catch (Throwable t) {
-                    cancelled.set(false);
-                    throw SMBRuntimeException.Wrapper.wrap(t);
-                } finally {
-                    lock.writeLock().unlock();
-                }
-            }
-
-            @Override
-            public boolean isCancelled() {
-                lock.readLock().lock();
-                try {
-                    return cancelled.get();
-                } finally {
-                    lock.readLock().unlock();
-                }
-            }
-
-            @Override
-            public boolean isDone() {
-                lock.readLock().lock();
-                try {
-                    return cancelled.get() || promise.isDelivered();
-                } finally {
-                    lock.readLock().unlock();
-                }
-            }
-
-            @Override
-            public T get() throws InterruptedException, ExecutionException {
-                logger.debug("Retrieving value for Future << {} >>", messageId);
-                //noinspection unchecked
-                return (T) promise.retrieve();
-            }
-
-            @Override
-            public T get(long timeout, TimeUnit unit) throws InterruptedException, ExecutionException, TimeoutException {
-                //noinspection unchecked
-                return (T) promise.retrieve(timeout, unit);
-            }
-        };
     }
 
     UUID getCancelId() {
@@ -135,7 +75,4 @@ class Request {
         return timestamp;
     }
 
-    interface CancelCallback {
-        void cancel(long messageId);
-    }
 }
